Many changes to get OSS 5.0 working
Added VersionConfig to track differences between versions easily
  and update params based on versions so the topo managers can
  adjust as necessary to differet version's available features.

Added UnitTestRUnner to prepare for running in jenkins jobs.  Now
  JUnit XML files can be produced just by adding an env var.

Various changes to adjust to new OSS 5.0 model (ZOOM + cluster).

Updates to tests to use the new UnitTestRunner and to work with
  new changes to the PTM/VTM for OSS 5.0

Changes to TSM to work in more of a scenario-based manner.  Now
  tests are added to a run map based on the scenarios they will
  run against (filtered by optional params).  Now a scenario will
  start only once, run ALL test cases and then stop.

Added env set up scripts and unit test run script to a new "bin"
  directory.

Change-Id: I0cdeb1eb26f84bbcca4ef69fe41598aa6c6c9c68