Fix echo server and ping issues
Fix the echo server back to using a  sub process, so it can start
in an IP netns.  This time, thought, keep the echo script simple
and easy to terminate, as well as provide a script to do the echo
sends, instead of using nc.

Add a verify command to verify connectivity between two hosts, but
be pretty relaxed about dropped packets at the start.  The former
commands will stay strict and fail if a packet is dropped, but we
need a way to test connections between hosts which can allow for
a dropped packet or so to get around flaky timing issues (mostly
due to topology caching on the first packet).

Update all tests which were using ping to use the verify command
instead (use_tcp=False for ping-only, and use_icmp=False when the
check was for TCP only).

Change-Id: I96a4c1f30a6c04c803bf9897178c5c71c17e6eeb