More fixes for missing pings
Added some test fixes to help assuage missing and timing out
pings.

Make sure to ping 127.0.0.1 for connection test for
echo-server if the designated IP is "" (all IPs) and use the
designated IP if it is specified and not empty (as 127.0.0.1
will fail in that case).

Change echo server to use Template Method pattern and do log +
checking in the base class method.

Change-Id: I79c9eba1304b9850b59a09974f27f1961b610224