Smooth out resource cleanup error-handling
When a resource cleanup fails, instead catalog all the failures
as they go by, and return them.  And for each function, keep a
running list of all the resource cleanup errors that occured,
so at the very end of the run, we can fail with a single test
failure with all the cleanup failures listed out.

Change-Id: I2b89f7b39f6d316cbd7fc898c912fc0f870a93ce