Use a proxy host to proxy a host to the root
The RootHost is a standard "on-root" host, but the hostname
itself is referred to by its config name (zoo1, net1, etc.).  For
some apps, it needs to actually know the real host name (returned
byt he 'hostname' command) in order to function, but since other
apps, wiring, etc. all need to use the host's config name, we
need to 'proxy' this host as the root host name, but keep the
configured name for configuration.

Change-Id: Ia298dbfccc2ac5b29ad0df345177be305795fcc1