Allow echo and ping to "prime" topology
Depending on the overlay (as in midonet's case), a list of
settings can be used to help govern zephyr behavior.  Added
a mechanism for the hosts to generate and keep a map of settings
to values, depending on the overlay applications which are
running (or were started in the PTM).

Added a "pre-caching-required" flag which signals to the echo
and ping functions that they should send a test packet to the
destination IP first, and then send the real packet.  If the first
packet fails, no problem, as it is just there to make sure the
overlay laods the topology info it needs.

Change-Id: Ie30a82fc1f7c4d0a61d01aa6f79d260178fe20fc