Allow net_id to be none if create_vm is given port
If create_vm_server is given an existing neutron port, there is
no need for a neutron net_id, as a port will not be created.
In this case, net_id can be None.  If the port is also not given,
an exception will be thrown.  So, eitehr one or the other needs
to be present (if both, the port will take precendence).

Change-Id: I4496639cc160a9026c75b61e608b046e3f988f4d