PTM-CTL and utils will now use existing underlay
If the underlay-config.json file is present (as it should be for
a running PTM, and it should be absent if no PTM is running), use
it to get the config information isntead of forcing users to
always supply the config (which they might have forgotten or is
annoying to type, etc.), since we have that information already
available.

Also flesh out the ptm-ctl usage text as well as how zephyr
configures its own root directories.

Change-Id: I107718b846994ffd89177ba2418163f185556423