Update FWaaS logging tests for binary logging
Update Midolman.get_resource() to read from either the legacy text firewall logs
or the new binary firewall logs, and return a list of dictionaries (one
dictionary per log entry) instead of plain text. If no UUID is specified, it
returns a dictionary of lists of dictionaries, with the outer dictionary mapping
the logger ID to the list of entries for that logger. This will allow more
detailed verification of the log contents.

Change-Id: Ic2ec0c98de9b07d895048ccdb6ac6744235af11a
Signed-off-by: Brandon Berg <bberg@midokura.com>