Change PTM's config file to file name only
The PTM's config file attribute was getting the full path in
the parameter, meaning that any check of the topology feature
was also checking against the full path, which is different on
different systems, so the check would never succeed.  Instead,
we should only check the config_file feature against the base
name of the file, which will discard system paths and only check
that the desired topo file was used even if the path was different.

Change-Id: Iba071153cffdd8f671d2a2ee74e828b8c501206e